requirements.txt
selene>=2.0.0rc4
pytest~=7.4.0
allure-pytest
selenium~=4.10.0
python-dotenv~=1.0.0